requirements.txt
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
fastapi==0.95.1
uvicorn==0.22.0
slack-sdk==3.26.1
python-dotenv==1.0.0
boto3>=1.34.0
email-validator>=2.1.0
pytest>=8.0.0
httpx>=0.26.0
jinja2==3.1.2
pyyaml==6.0.1
slowapi==0.1.8
python-multipart==0.0.6
bcrypt==4.0.1
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==11.0.3
psutil==5.9.5
scapy==2.5.0
netifaces==0.11.0
aiofiles==23.1.0
python-socketio==5.8.0
requests==2.31.0
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
aiosqlite==0.19.0